requirements.txt
line-bot-sdk==3.14.3
Flask==2.3.2
python-dotenv==1.0.0
openai==0.28.0
google-cloud-texttospeech==2.15.0
tinydb==4.7.1
BetterJSONStorage==1.3.1